feat: init the repository (#2)
# Description

Set up a repository with a dummy Maven project and workflows pushing to
Maven Central. For regular updates a script is provided, which
automatically creates a PR.

It also provides init branches for the license and codeowners file. If
these branches are present, a PR is automatically created. It contains
further instructions for the correct setup.

# Verification

Manually verified on the `test` repository.