feat: add code of conduct (#8)
# Description

This pull request introduces a code_of_conduct.md. The addition of this
file is crucial for several reasons:
- It sets clear expectations for behavior for all participants in the
project, fostering a positive and respectful environment.
- It provides a framework for handling and resolving behavioral issues,
including a process for reporting and addressing unacceptable behavior.
-It promotes an open, inclusive, and diverse community, which can help
attract a wider range of contributors.
- It helps establish trust with the community, showing that project
maintainers take community issues seriously and are committed to
resolving them.

# Verification

Not necessary. Just adding a text document.